Move the inner evaluation functions of machine to eval submodule
